# src/routes/openapi/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BlockchainNetwork": {
				"description": "Supported blockchain networks",
				"enum": [
					"ethereum",
					"bsc",
					"polygon",
					"solana",
					"arbitrum",
					"optimism",
					"avalanche",
					"base"
				],
				"type": "string"
			},
			"TokenInfo": {
				"description": "Token info interface for tracking specific tokens",
				"properties": {
					"contractAddress": {
						"type": "string"
					},
					"network": {
						"$ref": "#/components/schemas/BlockchainNetwork"
					},
					"symbol": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"contractAddress",
					"network",
					"symbol",
					"name",
					"decimals"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTokenBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"network": {
						"type": "string"
					},
					"contractAddress": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"symbol",
					"network"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateTokenBody_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"network": {
						"type": "string"
					},
					"contractAddress": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CreateReminderRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"reminderTime": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"reminderTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PortfolioChartDataPoint": {
				"description": "Interface for chart data point",
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"totalValueUsd": {
						"type": "number",
						"format": "double"
					},
					"walletsValueUsd": {
						"type": "number",
						"format": "double"
					},
					"exchangeValueUsd": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"timestamp",
					"totalValueUsd",
					"walletsValueUsd",
					"exchangeValueUsd"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "your-node-app-name",
		"version": "1.0.0",
		"description": "Describe your app here.",
		"contact": {
			"name": "author_name"
		}
	},
	"paths": {
...
		"/tokens": {
			"get": {
				"operationId": "GetAllTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TokenInfo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTokenBody"
							}
						}
					}
				}
			}
		},
		"/tokens/network/{network}": {
			"get": {
				"operationId": "GetTokensByNetwork",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TokenInfo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "network",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tokens/{id}": {
			"put": {
				"operationId": "UpdateToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_CreateTokenBody_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reminders": {
			"post": {
				"operationId": "CreateReminder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"title": {
											"type": "string"
										},
										"reminderTime": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"completed": {
											"type": "boolean"
										}
									},
									"required": [
										"updatedAt",
										"title",
										"reminderTime",
										"id",
										"description",
										"createdAt",
										"completed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reminders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReminderRequest"
							}
						}
					}
				}
			}
		},
...
		"/reminders/{reminderId}/complete": {
			"put": {
				"operationId": "UpdateReminderCompletion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"title": {
											"type": "string"
										},
										"reminderTime": {
											"type": "string",
											"format": "date-time"
										},
										"id": {
											"type": "string"
										},
										"description": {
											"type": "string"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"completed": {
											"type": "boolean"
										}
									},
									"required": [
										"updatedAt",
										"title",
										"reminderTime",
										"id",
										"description",
										"createdAt",
										"completed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reminders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "reminderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reminders/{reminderId}": {
			"delete": {
				"operationId": "DeleteReminder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Reminders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "reminderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/portfolio/chart-data": {
			"get": {
				"operationId": "GetChartData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/PortfolioChartDataPoint"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"message",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Portfolio"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "days",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}